+++ tests/TestCase.php
@@ -3,11 +3,12 @@
 namespace Sfneal\Helpers\Redis\Tests;
 
 use Illuminate\Foundation\Application;
+use Illuminate\Foundation\Bootstrap\LoadEnvironmentVariables;
 use Lunaweb\RedisMock\Providers\RedisMockServiceProvider;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
 use Sfneal\Helpers\Redis\Providers\RedisHelpersServiceProvider;
 
-class TestCase extends OrchestraTestCase
+abstract class TestCase extends OrchestraTestCase
 {
     /**
      * Define environment setup.
@@ -17,8 +18,13 @@ class TestCase extends OrchestraTestCase
      */
     protected function getEnvironmentSetUp($app)
     {
+        parent::getEnvironmentSetUp($app);
+
+        // make sure, our .env file is loaded
+        $app->useEnvironmentPath(__DIR__.'/..');
+        $app->bootstrapWith([LoadEnvironmentVariables::class]);
+
         $app['config']->set('app.debug', true);
-        $app['config']->set('database.redis.client', 'mock');
         $app['config']->set('cache.default', 'redis');
         $app['config']->set('cache.prefix', 'redis-helpers');
     }