+++ tests/Unit/Traits/ExistsTest.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Sfneal\Helpers\Redis\Tests\Unit\Traits;
+
+use Sfneal\Helpers\Redis\RedisCache;
+
+trait ExistsTest
+{
+    /**
+     * @test
+     */
+    public function key_exists()
+    {
+        $key = 'bos-14';
+        $value = 'w';
+        RedisCache::set($key, $value);
+        $actual = RedisCache::exists($key);
+
+        $this->assertTrue($actual);
+    }
+}